refactor: Move uma_protection behavior to separate class
