+++ packages/uma/config/routes/tokens.json
@@ -10,9 +10,13 @@
       "handler": {
         "@type": "TokenRequestHandler",
         "negotiator": { "@id": "urn:uma:default:Negotiator" },
-        "storage": { "@id": "urn:solid-server:default:ClientRegistrationStorage" },
-        "keyGen": { "@id": "urn:uma:default:JwkGenerator" },
-        "baseUrl": { "@id": "urn:uma:variables:baseUrl" }
+        "umaProtection": {
+          "@id": "urn:uma:default:UmaProtection",
+          "@type": "UmaProtection",
+          "storage": { "@id": "urn:solid-server:default:ClientRegistrationStorage" },
+          "keyGen": { "@id": "urn:uma:default:JwkGenerator" },
+          "baseUrl": { "@id": "urn:uma:variables:baseUrl" }
+        }
       },
       "path": "/uma/token"
     }
